+++ Parley/Parley/Views/Helpers/FileMenuController.cs
@@ -8,8 +8,8 @@ using Avalonia.Interactivity;
 using Avalonia.Platform.Storage;
 using DialogEditor.Parsers;
 using DialogEditor.Services;
+using Radoub.Formats.Common;
 using Radoub.Formats.Logging;
-using DialogEditor.Utils;
 using DialogEditor.ViewModels;
 
 namespace Parley.Views.Helpers